Updating main
I jumped to part 10 and decided to make a function to compare the input passed in against the built-in functions- exit and cd. I set up the exit and cd function (just so they could be jumped to when called within the stringCompare function) and will modify their functionality soon.